Change id keys to bigInteger to account for Date.now()

+++ db/migrations/20210210092817_conversations.js
@@ -1,7 +1,7 @@
 exports.up = function(knex) {
   return knex.schema
     .createTable('conversations', function (table) {
-      table.integer('conversation_id').primary();
+      table.bigInteger('conversation_id').primary();
       table.string('title');
       table.date('start_date').defaultTo(knex.fn.now());
     })
