+++ db/migrations/20210210093720_messages.js
@@ -1,11 +1,11 @@
 exports.up = function(knex) {
   return knex.schema
     .createTable('messages', function (table) {
-      table.integer('message_id').primary();
+      table.bigInteger('message_id').primary();
       table.string('text');
       table.date('date_sent').defaultTo(knex.fn.now());
       table.time('time_sent');
-      table.integer('conversation_id').notNullable()
+      table.bigInteger('conversation_id').notNullable()
         .references('conversation_id').inTable('conversations').onDelete('cascade');
     })
 };
